Added blocking for clients

+++ core/src/main/scala/com/raphtory/internals/components/querymanager/AnalysisMessages.scala
@@ -183,4 +183,4 @@ private[raphtory] case class ClientDisconnected(graphID: String, clientID: Strin
 
 case class BlockIngestion(sourceID: Int, graphID: String) extends QueryManagement
 
-case class UnblockIngestion(sourceID: Int, graphID: String, messageCount: Int, force: Boolean) extends QueryManagement
+case class UnblockIngestion(sourceID: Int, graphID: String, messageCount: Long, force: Boolean) extends QueryManagement
